requirements.txt
fastapi
uvicorn
openai-whisper
pyannote-audio
spacy
python-multipart
